requirements.txt
apache-airflow==1.10.10
pytest-airflow==0.0.3
Jinja2==3.1.2